Implement entrypoints for pub IIB methods
The new entrypoints are meant to be used by the pub methods
"PushAddIIBBundles", "PushRemoveIIBOperators", and
"PushIIBBuildFromScratch" and are invoked similiarly as push-docker.
To avoid code duplication, several methods were turned to class
methods so they could be invoked independently from the object's
context. Namely, these are: "iib_add_bundles", "iib_remove_operators",
and "run_tag_images". Image signing workflow unique to these pub
tasks was implemented in "sign_task_index_image".

One original design decision has changed by this task: it was decided
that pubtools-iib will not invoke pubtools-quay for pushing the new
index image to quay. Instead, the logic for calling pubtools-iib,
and subsequently pushing the image to its destination will be handled
in pubtools-quay. This was decided to avoid increasing the complexity
of pubtools-iib, and to simplify the implementation on pub's side.

Side effect is that the full operator workflow cannot be invoked as
a CLI script. The reason for this decision is that the original
design hasn't accounted for index image signing, which wouldn't be
performed by a CLI script anyway. The current technical solution
allows for an extension of CLI capabilities, but the argument input
would be complex and the script would still need to contact pub-hub
for the signing operation.