Fix signing to use existing index image
In signing, index image is primarily used for getting its manifest
list. As the destination index image doesn't exist at the moment of
signing, image built by IIB has to be used. Customer-facing image
name is derived from target settings, with one exception: image tag.
Tag was derived from the provided index image, which poses no issues
for "add bundles" and "remove operators", as the tags of image built
by IIB and image served to customers are the same. It causes issues
for "build from scratch", as the desired tag is specified by the
user. For this reason, an option to specify a custom tag for signing
was added, which will override the tag derived from the index image.