Use intermediate index image for signing
IIB builds all images to the intermediate "iib" repo. From there,
they are pushed to the desired location. When "overwrite_from_index"
is specified, the destination image is overwritten. When signing the
new index image, destination reference is used to get the manifest
list containing new digests to be signed. However, a scenario may
occur where the destination is overwritten by a new index image
("overwrite_from_index" is True), and correct digests cannot be
retrieved. This is why a change is made to use the intermediate
index image for gathering the manifest list, which has no risk of
being overwritten.

A question may arise: "why are we signing something that is no longer
available to customers?" Well, I don't know.