refactor: use hooks to declare events of interest
pubtools library includes a hook/event system. Use that to
declare hooks which may be of interest to Pub.

The primary motivation for this right now is to address some
problems with the UMB message sending, hence the addition of one
hook nearby each point where a UMB message could be sent.

The idea is to do the following:

1. add these hooks in pubtools-quay
2. make Pub connect to these hooks and send UMB messages
3. drop the UMB sending code from pubtools-quay

The main reason for this is that pubtools-quay simply doesn't have
access to all the needed information in order to send appropriate
messages. For example, it does not know a Pub task ID, a Pub target or
the name of the user who requested the task, all of which should be
included.

Another reason is that each project should not have to independently
deal with the necessary logic for connecting to UMB and sending
messages. As can be seen in pubtools-quay, the amount of code needed to
do that (and test it) is far from trivial.