ci: drop python 2.6 configuration
There are currently some issues with the python 2.6 test configuration.

They've been triggered by the release of pubtools-1.1.0, which requires
that all installed pubtools-* entry points are importable; but this is
not true on the python 2.6 environment set up here, because some of the
installed library versions aren't compatible with python 2.6.

In some cases this isn't easy to resolve, because internal packages
have been made py2.6-compatible via patches applied during RPM build,
and similar patched versions aren't available anywhere in a
pip-installable form. "attrs" library is an example of this - there's
no upstream version of it which is both py2.6 compatible and also has
all the API used by pubtools-pulplib.

I propose dropping the python 2.6 configuration as a solution to this.
Most of our other github projects already don't test python 2.6 in
their CIs. If you compare the cost of maintaining it against the benefit
you get back, I don't think it's worthwhile:

- It's more complicated & slower than other CI configs

- Differences between python 2.6 and 2.7 are relatively minor (certainly
  compared to py2 vs py3). It's just not all that likely to make changes
  which work on py2.7 and fail on py2.6.

- In the case where that *does* happen: we have another layer of
  integration testing beyond this repo's CI anyway, which uses a more
  accurate environment (RHEL6). It's reasonable to leave py2.6
  compatibility within the scope of that later testing.

- Many of the advantages of testing py2.6 in CI can be gained by
  requesting older versions of libraries within the 2.7 test config;
  i.e. investigating which version of a library will be used on the
  production py2.6 environment and pinning to the same version here.
  This at least allows verifying compatibility with the libraries we
  expect to use on py2.6.  This approach has been included within the
  commit, adding a constraints file with old versions of libraries.