Remove skopeo image building data and executor image authentication
Container image used for running skopeo operations requires Redhat
CA certificate to work properly. However, this certificate should
not be shared publicly, so it has to be removed from this repo.
For this reason, Dockerfile definition, build, and push can no longer
be done with the current workflow. Instead, container image definition
will be moved to a private Gitlab repo, and its public Quay image
will be removed.

The image will be stored in a private Quay repository. As credentials
will be needed to pull this image, option to log in to a registry
has been added to ContainerExecutor. The credentials are expected
to be gathered from target settings.