Make IIB tag new index images with unique tag and use it to get manifest
A while ago it was discovered that using floating tag of index images
to get manifest list may return the incorrect result due to a race
condition (index image was overwritten by a parallel push). The solution
was to use intermediate index image specified by digest. However, as
this index image may be overwritten as well, the new solution is to
make IIB tag the new index image with a unique tag and use it to
get the manifest list for signing. Tag uniqueness is ensured by
combining the index image floating tag (OCP version) and pub task
ID. This change was applied to all IIB calls done by pubtools-quay,
so standard push-docker workflow and separate IIB-only workflows.

Few minor unrelated changes were done as well, namely cleanup of
unused variables, docstring fixes, and change of "failed" variable
in push_docker.py to bool instead of integer because this isn't C.