Make Pyxis repo check less prone to failure
It was discovered that some repositories don't have metadata in
Pyxis at all, so 404 error should be ignored. In addition, some
repo metadata may be missing the "release_categories" field, so
the check will not fail if the field is missing. Lastly, deprecated
repos are tolerated by default, since the check would cause
issues during data migration. Based on the recent issues, the presence
of "Deprecated" flag may not be as reliable as we'd like, so
the check will be disabled until we get more confidence on its
reliability. The consequences of pushing to deprecated repos are
basically non-existent, so it shouldn't cause any issues.