Sign V2S1 manifests
Since Quay.io generates V2S1 manifest for an image on demand,
signatures for digests of these manifests have to be created
as well. As the manifest values depend on the destination
location, the V2S1 manifests can only be queried after the
images are pushed. For this reason, signing of V2S1 digests
is done separately, after the image push. An unfortunate
side effect is that there will be a small time window where
a customer may pull a newly pushed image whose signature has
not been generated yet.

Multiarch images are not relevant for V2S1 because it predates
the existence of a manifest list. If the signing of V2S1 digests
fails, some newly pushed content may be left unsigned
indefinitely. I decided against implementing a rollback
functionality, since even rollback is only best-effort, and in
case of its failure nothing else guarantees the validity
of pushed data. If a push fails, the issues should be
resolved by a retried push.

The logic to remove no longer used signatures was extended for
V2S1 digests. In case of overwriting an existing tag, its
schema1 signatures will be removed as well.

A few unrelated things were fixed in this commit:
- "operator_push_items" parameter was removed from
  "remove_old_signatures" method because it wasn't used
  anywhere
- "esig" variable in "remove_old_signatures" was renamed
  to "existing_signature"
- "failed" variable in "run" is now a boolean instead of
  using 0/1 for True/False