Remove the option for entrypoints to send UMB messages
This capability was initially implemented due to a misunderstanding
with the Metaxor team, which captures UMB messages to know what
metadata to refresh. Pub creates a UMB message for each new task,
which is what Metaxor currently captures. This means that it's
unnecessary to generate these messages by the entrypoints themselves,
as they share no real purpose and are not consumed by anyone. Removing
this functionality simplifies the code and removes a requirement to
install an internal dependency.