Unpin the version of python-qpid-proton in pubtools-quay installation
The version of python-qpid-proton was originally pinned to 0.33.0 so
that the Python 2 test suite would pass. However, this constraint has no
reason to exist during Python 3 installation, and causes issues when
pubtools-quay is invoked as a console script. Move the version
constraint to requirements-test.txt and only apply it to Python 2.