Change condition to not require hashing
The condition checking for IIB failures fails on Python 3, since it
attempts to hash an unhashable object. This happens during an attempt to
create a set from a list containing a mix of iiblib models and Falses.
IIBlib models are appearently not hashable on Python 3. Change the
condition to not require hashing.