Poll for consistent results of whether a tag exists
It's possible that Docker HTTP API will claim that a tag exists, but
trying to get digest of this tag results in a 404. It can't be
determined which source is right, so poll both until the results are
consistent. If the tag ends up not existing, add it to rollback tags,
and if it begins to exist continue with the backup tags workflow. The
polling will not happen indefinitely, at some point the method will give
up and assume that the tag doesn't exist.