Push multiarch image when the current destination doesn't have a ML
When pushing a multiarch image, the current workflow expects the
destination to either be empty or contain a multiarch image. If latter,
merging of manifest lists is attempted. If the destination image doesn't
contain a ML, an error is raised. Modify the code to be able to handle a
scenario when the destination is not a multiarch image. The existing
manifest will be overwritten by the manifest list without attempting a
merge.