Use a random filename for the password file in containers
Bandit has detected a possible vulnerability of using a /tmp directory
for injecting a password file. Changing the used directory seems to have
caused some unexpected issues, so revert back to using /tmp and add a
random suffix to the injected file so that an attacker couldn't inject
their own value.