Remove less signatures [CLOUDDST-19242] (#173)
* Remove less signatures [CLOUDDST-19242]

If signatures of a reference were partially uploaded due to an
issue, re-push would add signatures of 'missing digests', but take
the previously uploaded ones as outdated and remove them.

This patch excludes 'missing digests' signatures from new to avoid
over removal. It's not ideal because unused signatures might not
be removed in some cases. Until outdated signatures can be identified
in a more accurate way, removing less is better than overly.

* filter out signatures to remove based on rollback_tags

* Updated condition to skip removal

* Fixed condition for old signatures removal

* Fixed tests

---------

Co-authored-by: Jindrich Luza <midnightercz@gmail.com>
Co-authored-by: Jindrich Luza <jluza@redhat.com>