Implement workflow to push container security manifests
The workflow can be described as follows: Get security manifest for each
architecture of a given push item. Then check if an attestation of this
archtecture already exists in the destination. If it does, determine if
a new product name will be added to the properties. If it does,
overwrite the attestation with the new product name included. If it
doesn't, skip uploading the attestation, as no new information will be
added. If an attestation doesn't exist in the destination, just upload
the security manifest.

The workflow gets the product name from the push item, which gets it
from Errata via pushsource. PushDocker method doesn't have an associated
advisory, so there is not enough information to get the associated
product. The workflow supports not adding product name to the security
manifest in such cases.

Only multiarch push items and source push items are supported. Other
image types (for example v2s1) are assumed to not contain a security
manifest and are skipped. In fact, many source and multiarch images
don't contain security manifests either, which is why no error is
raised and the push continues without uploading them.

Cosign has several limitations:
- It can only generate attestation images to images that are already
  pushed to the destination. This means that if the push fails after an
  image is pushed, but before the attestation image is uploaded, an
  image will exist without a security manifest. This was deemed
  acceptable since the affected tags won't be visible in the catalog
  because Metaxor only processes sucessful pub tasks.
- It can only use the default docker auth config file location to get
  credentials. Pub currently has two config files. The file in the
  default location (/.docker/config.json) contains pre-filled data and
  is immutable because it's mounted from a Kubernetes resource. The
  second config file is only used by skopeo and its location is
  specified with the REGISTRY_AUTH_FILE env variable. Skopeo login
  writes the acquired credentials to this file at runtime. The issue for
  cosign is that the destination credentials (needed to get and push
  attestations) are target-specific, and thus can only be set at
  runtime. It is resolved by making the default config file
  writable (mounting the Kubernetes resource to a different file and
  copying the contents to /.docker/config.json at startup) and by
  invoking skopeo login at the start of the security manifest pushing
  stage.
- It cannot easily overwrite existing attestations. When attempting to
  push a security manifest when one was already pushed, the resulting
  attestation will containg two security manifests, which is obviously
  incorrect. If attempting to use the --replace flag, the command fails.
  To resolve it, the existing attestation is maually removed before the
  new one is pushed. This may pose issues if a push fails between
  removing the old and pushing the new attestations, since a published
  image may now not have a security manifest. In the following section I
  will explain why I consider this acceptable.

It is very unlikely that a container image can be associated with
multiple products. This means that the workflow where an existing
attestation image is removed and merged will almost never occur.
Instead, attestation images will simply be copied to the destination.