Add static typing to pubtools-quay
The mypy check succeeds with strict flag enabled. Some unused files are
removed as a part of this commit, so that static typing didn't have to
be added to them. Some code has been moved to a higher location so that
the type it creates is available to subsequent functions. A few comments
have been added with the TOFIX prefix. They describe unusual behavior
that may be a mistake in the code. These mistakes were found in the new
signing related code and should be investigated.