Generate SBOM attestations for manifest lists
When attempting to get SBOMs of multiarch images referenced via tag,
cosign returns no results. The reason is that it calculates the digest
of the manifest list, which has no associated SBOM images. This can be
solved by also generating attestations for the manifest list digests.

There are no guidelines of what a ML SBOM should contain, but there is
precedent of it containing either a merge of all arch image packages, or
all arch SBOMs in a JSONL format. The second option was chosen for
implementation, which can be achieved by calling cosign attest on the
same digest multiple times.

The arch attestations that should be a part of ML attestation are
determined by inspecting the arch digests of a given ML. If a ML
attestation already exists, it's always removed and replaced by a new
one, even if there are no meaningful changes.

I believe that no changes are necessary for untagging images, since ML
digest is already added to the lost digests.