Raise an error when manifest claims retry limit is reached
ManifestClaimsHandler automatically retries when a message encounters an
error. However, when a retry limit is reached, the handler doesn't raise
an error and simply continues the push with missing messages. This is
obviously an issue as it results in some signatures not being created in
the push. Fix the issue by raising an error when the retry limit is
reached.

NOTE: At the time of writing this fix, the affected code is no longer
used in production, and is superseded by pubtools-sign. The fix is
implemented only if a rollback to the old code was necessary.