Ensure that unusual cosign errors are raised
Originally, getting a nonzero code from "cosign download sbom" or
"cosign verify-attestation" was interpreted as SBOM/attestation not
existing. This isn't necessarily the case, as the commands may fail for
other reasons (for example networking issues). The code assumes that
existing attestation doesn't actually exist, which can cause unexpected
behavior like the duplication of the attestation data.

Fix the issue by raising any unexpected errors. To minimize transient
errors, add retry mechanism to the cosign commands. If the command fails,
it is retried multiple times before the task is failed.